% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{add_visit_count}
\alias{add_visit_count}
\title{Add a visit counter to the raw data file.}
\usage{
add_visit_count(data, vars = c("SID", "VisitDate"), start_count = 1)
}
\arguments{
\item{data}{Processing data.}

\item{vars}{Variable to add visit information by (ie. SID with
earlier VisitDate would be 1, increasing up as VisitDates increase).}

\item{start_count}{When to start the counting (ie. if for first visit, it
would be 1)}
}
\description{
Add a visit counter to the raw data file.
}
