% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining.R
\name{combine_datasets}
\alias{combine_datasets}
\title{Combine multiple datasets together into one, merged dataset. Merges based on
SID and VN columns.}
\usage{
combine_datasets(..., by_cols = c("SID", "VN"))
}
\arguments{
\item{...}{Any number of datasets that need to be merged.}

\item{by_cols}{Combine datasets based on these variables.}
}
\description{
Combine multiple datasets together into one, merged dataset. Merges based on
SID and VN columns.
}

