% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_date}
\alias{convert_to_date}
\title{Convert a date variable into another format (e.g. ISO standard).}
\usage{
convert_to_date(variable, from = "\%m/\%d/\%y")
}
\arguments{
\item{variable}{The date variable.}

\item{from}{Original date format.}
}
\description{
Convert a date variable into another format (e.g. ISO standard).
}
