% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{drop_empty_rows}
\alias{drop_empty_rows}
\title{Drops rows that are completely empty (excluding SID, etc).}
\usage{
drop_empty_rows(data, col_nums)
}
\arguments{
\item{data}{The processing dataset.}

\item{col_nums}{Numeric value for the columns to exclude or include (e.g.
-1:-2 to exclude the first two columns, or 3:10 to include columns 3 to
10).}
}
\description{
Drops rows that are completely empty (excluding SID, etc).
}
