% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{drop_na_variables}
\alias{drop_na_variables}
\title{Drop the variables named 'NA' from the raw dataset.}
\usage{
drop_na_variables(data)
}
\arguments{
\item{data}{The processing dataset.}
}
\description{
Drop the variables named 'NA' from the raw dataset.
}

