% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_final.R
\name{export_data}
\alias{export_data}
\title{Export a dataset into another format (Rds, csv, or as a package dataset).}
\usage{
export_data(data, name = deparse(substitute(data)), type = c("Rds",
  "csv", "pkg"), path = getOption("PROMISE.extdata.path"))
}
\arguments{
\item{data}{The dataset.}

\item{name}{The name of the output dataset filename (without filename
extensions).}

\item{type}{The format to output/save.}

\item{path}{The path to save the dataset to.}
}
\description{
Export a dataset into another format (Rds, csv, or as a package dataset).
}
