% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{extract_vn}
\alias{extract_vn}
\title{Extract the VN information from the Visit column (e.g. that has the data as
'yr3', or 'YR6', or 'y1', etc.).}
\usage{
extract_vn(data, original_visit = "Visit")
}
\arguments{
\item{data}{Processing data.}

\item{original_visit}{Visit date variable to extract the VN from.}
}
\description{
Extract the VN information from the Visit column (e.g. that has the data as
'yr3', or 'YR6', or 'y1', etc.).
}

