% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{fix_date}
\alias{fix_date}
\title{Fix with with any date variable.}
\usage{
fix_date(data, date_var, original_format = "\%m/\%d/\%y")
}
\arguments{
\item{data}{The processing dataset.}

\item{date_var}{The date variable found in the dataset.}

\item{original_format}{The original format of the date (e.g. MM/DD/YY is
\%m/\%d/\%y or YYYY-MM-DD is \%Y-\%m-\%d).}
}
\description{
Fix with with any date variable.
}
\seealso{
To see the different options for date formats, see
  \code{\link[base]{strptime}}.
}
