% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_csv}
\alias{import_csv}
\title{General function; import raw csv datafile into R.}
\usage{
import_csv(file, path = NULL, skip_lines = 0)
}
\arguments{
\item{file}{The name of the raw csv datafile that is found in the
\code{data-raw/} folder.}

\item{path}{Path to file.}

\item{skip_lines}{Whether to skip n lines when importing the csv file.}
}
\description{
General function; import raw csv datafile into R.
}

