% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_csv_multiple}
\alias{import_csv_multiple}
\title{General function; imports multiple raw data
files based on a pattern, only used for data of the same type (e.g. 'ogtt' or
'fattyacids').}
\usage{
import_csv_multiple(pattern, merge_direction = c("rows", "columns"),
  path = getOption("PROMISE.data.raw.path"))
}
\arguments{
\item{pattern}{The file pattern to search for in the
\code{data-raw/} folder.}

\item{merge_direction}{Whether to combine datasets by rows or by columns.
This is highly situational... Mostly will need by rows.}

\item{path}{Path to directory where the files are stored.}
}
\description{
General function; imports multiple raw data
files based on a pattern, only used for data of the same type (e.g. 'ogtt' or
'fattyacids').
}

