% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiple_files}
\alias{multiple_files}
\title{Get a vector of files with the full path for a given regular expression.}
\usage{
multiple_files(pattern, path)
}
\arguments{
\item{pattern}{Regular expression pattern to search for files.}

\item{path}{Where to search for the files.}
}
\description{
Get a vector of files with the full path for a given regular expression.
}

