% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{rename_variables}
\alias{rename_variables}
\title{Rename the raw dataset variable names based on the yaml files.}
\usage{
rename_variables(.data, yaml_file, path = NULL)
}
\arguments{
\item{.data}{The processing dataset.}

\item{yaml_file}{The yaml renaming files in the \code{inst/rename/} folder.}

\item{path}{Path to the yaml file. Defaults to the rename folder.}
}
\description{
Rename the raw dataset variable names based on the yaml files.
}
