% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_duplicates.R
\name{scr_duplicates}
\alias{scr_duplicates}
\title{Fix any duplicates by dropping them.}
\usage{
scr_duplicates(.data, .variables, action = c("keeplast", "keepfirst"))
}
\arguments{
\item{.data}{Data with duplicates.}

\item{.variables}{Variables (columns) to use as the basis for determining duplicates.}

\item{action}{The action to take for duplicates:\preformatted{- "keeplast": Keep the last duplicate value.
- "keepfirst": Keep the first duplicate value.
}}
}
\value{
Data without duplicates from specific columns.
}
\description{
Fix any duplicates by dropping them.
}
\examples{

ds <- data.frame(
a = c(rep(1, 3), rep(2, 3)),
b = c(1:3, 1, 1, 2),
v = rnorm(6)
)
ds \%>\%
scr_duplicates(c("a", "b"), "keepfirst")
}
