% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{spread_over_visits}
\alias{spread_over_visits}
\title{Spread a variable values from one visit date to the next.}
\usage{
spread_over_visits(.data, column)
}
\arguments{
\item{.data}{The processing dataset.}

\item{column}{Variables to spread values across over the visits.}
}
\description{
Spread a variable values from one visit date to the next.
}
