% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_variables.R
\name{average_variables}
\alias{average_variables}
\title{Average variables that have two or more measurements.}
\usage{
average_variables(data, var_pattern)
}
\arguments{
\item{data}{The scrubbing dataset.}

\item{var_pattern}{Regular expression pattern to search for variable names.}
}
\description{
Average variables that have two or more measurements.
}
